+++ src/configuration/applier/service.cc
@@ -360,7 +360,7 @@ void applier::service::modify_object(
            << host_name << "'");
 
   // Find service object.
-  umap<std::pair<unsigned int, unsigned int>,
+  umap<std::pair<unsigned long, unsigned long>,
        std::shared_ptr<service_struct> >::iterator
     it_obj(applier::state::instance().services_find(obj.key()));
   if (it_obj == applier::state::instance().services().end())
@@ -602,7 +602,7 @@ void applier::service::remove_object(
   // Find service.
   std::pair<std::string, std::string>
     id(std::make_pair(host_name, service_description));
-  umap<std::pair<unsigned int, unsigned int>,
+  umap<std::pair<unsigned long, unsigned long>,
        std::shared_ptr<service_struct> >::iterator
     it(applier::state::instance().services_find(obj.key()));
   if (it != applier::state::instance().services().end()) {
@@ -671,7 +671,7 @@ void applier::service::resolve_object(
     << "' of host '" << *obj.hosts().begin() << "'.";
 
   // Find service.
-  umap<std::pair<unsigned int, unsigned int>,
+  umap<std::pair<unsigned long, unsigned long>,
        std::shared_ptr<service_struct> >::iterator
     it(applier::state::instance().services_find(obj.key()));
   if (applier::state::instance().services().end() == it)
@@ -685,8 +685,8 @@ void applier::service::resolve_object(
     &deleter::objectlist);
 
   // Find host and adjust its counters.
-  unsigned int host_id(it->first.first);
-  umap<unsigned int, std::shared_ptr<com::centreon::engine::host>>::iterator
+  unsigned long host_id(it->first.first);
+  umap<unsigned long, std::shared_ptr<com::centreon::engine::host>>::iterator
     hst(applier::state::instance().hosts_find(it->first.first));
   if (hst != applier::state::instance().hosts().end()) {
     hst->second->set_total_services(hst->second->get_total_services() + 1);
